Refactor infrastructure to reduce references
