--- DAPM/DAPM.AccessControlService/Core/Dtos/PipelineDto.cs
@@ -1,10 +1,0 @@
-using DAPM.AccessControlService.Core.Domain.Entities;
-
-namespace DAPM.AccessControlService.Core.Dtos;
-
-public class PipelineDto
-{
-    public Guid Id { get; set; }
-    
-    public PipelineId ToPipelineId() => new(Id);
-}
