--- DAPM/DAPM.AccessControlService/Core/Dtos/RepositoryDto.cs
@@ -1,10 +1,0 @@
-using DAPM.AccessControlService.Core.Domain.Entities;
-
-namespace DAPM.AccessControlService.Core.Dtos;
-
-public record RepositoryDto
-{
-    public Guid Id { get; set; }
-    
-    public RepositoryId ToRepositoryId() => new(Id);
-}
