--- DAPM/DAPM.AccessControlService/Core/Dtos/UserDto.cs
@@ -1,10 +1,0 @@
-using DAPM.AccessControlService.Core.Domain.Entities;
-
-namespace DAPM.AccessControlService.Core.Dtos;
-
-public record UserDto
-{
-    public Guid Id { get; set; }
-    
-    public UserId ToUserId() => new(Id);
-}
