+++ DAPM/RabbitMQLibrary/Models/AccessControl/PipelineDto.cs
@@ -1,0 +1,7 @@
+namespace RabbitMQLibrary.Models.AccessControl;
+
+public class PipelineDto
+{
+    public Guid Id { get; set; }
+    
+}
