+++ DAPM/RabbitMQLibrary/Models/AccessControl/RepositoryDto.cs
@@ -1,0 +1,7 @@
+namespace RabbitMQLibrary.Models.AccessControl;
+
+public record RepositoryDto
+{
+    public Guid Id { get; set; }
+    
+}
