+++ DAPM/RabbitMQLibrary/Models/AccessControl/ResourceDto.cs
@@ -1,0 +1,7 @@
+namespace RabbitMQLibrary.Models.AccessControl;
+
+public record ResourceDto
+{
+    public Guid Id { get; set; }
+    
+}
