+++ DAPM/RabbitMQLibrary/Models/AccessControl/UserDto.cs
@@ -1,0 +1,6 @@
+namespace RabbitMQLibrary.Models.AccessControl;
+
+public record UserDto
+{
+    public Guid Id { get; set; }
+}
